+++ src/wormhole/__init__.py
@@ -1,7 +1,13 @@
 from ._rlcompleter import input_with_completion
 from .wormhole import create
+from ._status import WormholeStatus, DilationStatus  # export as public API
 
 from . import _version
 __version__ = _version.get_versions()['version']
 
-__all__ = ["create", "input_with_completion", "__version__"]
+__all__ = [
+    "__version__",
+    "create", "input_with_completion",
+    "WormholeStatus",
+    "DilationStatus",
+]