+++ src/wormhole/wormhole.py
@@ -88,6 +88,8 @@ class _DelegatedWormhole(object):
             raise NoKeyError()
         return derive_key(self._key, to_bytes(purpose), length)
 
+    # FFFFFUUUUk sakes, DelegatedWormhole doesn't support dilation??
+
     def close(self):
         self._boss.close()
 
@@ -123,7 +125,7 @@ class _DelegatedWormhole(object):
 
 @implementer(IWormhole, IDeferredWormhole)
 class _DeferredWormhole(object):
-    def __init__(self, reactor, eq, _enable_dilate=False):
+    def __init__(self, reactor, eq, _enable_dilate=False, _on_status_update=None):
         self._reactor = reactor
         self._welcome_observer = OneShotObserver(eq)
         self._code_observer = OneShotObserver(eq)
@@ -134,11 +136,19 @@ class _DeferredWormhole(object):
         self._received_observer = SequenceObserver(eq)
         self._closed = False
         self._closed_observer = OneShotObserver(eq)
+        # this will receive just "WormholeStatus" updates, where was
+        # the .dilate(..., on_status_update=) receives dilation
+        # updates (so effectively both)
+        self._on_status_update = _on_status_update  ##XXX is this a callable? or what?
 
         self._enable_dilate = _enable_dilate
 
     def _set_boss(self, boss):
         self._boss = boss
+        self._boss._on_status_update = self._on_status_update
+
+    # XXXX this is the right place for a "send_ping" or so, assuming that's useful
+    # roundtrip_ping() -> Deferred (callback() on Pong)
 
     # from above
     def get_code(self):
@@ -191,10 +201,27 @@ class _DeferredWormhole(object):
             raise NoKeyError()
         return derive_key(self._key, to_bytes(purpose), length)
 
-    def dilate(self, transit_relay_location=None, no_listen=False):
+    # XXX note that only DeferredWormhole has this .. the
+    # DelegatedWormhole thus has no way to dilate?
+
+    # XXX transit_relay_locations (plural) probably, and ability to
+    # pass a list? (there's a TODO about this is connector.py too)
+    def dilate(self, transit_relay_location=None, no_listen=False, on_status_update=None, ping_interval=None):
+        """
+        :returns EndpointRecord: an EndpointRecord containing the three
+            Twisted endpoint objects required to interact with the
+            Dilation channel (as control, connect and listen members).
+        """
         if not self._enable_dilate:
             raise NotImplementedError
-        return self._boss.dilate(transit_relay_location, no_listen)  # returns endpoints
+        return self._boss.dilate(transit_relay_location, no_listen, on_status_update, ping_interval)
+
+    def when_closed(self):
+        """
+        :returns: A Deferred that fires when this connection is closed
+            (calling this does _not_ initiate a close from our side)
+        """
+        return self._closed_observer.when_fired()
 
     def close(self):
         # fails with WormholeError unless we established a connection
@@ -226,6 +253,8 @@ class _DeferredWormhole(object):
         self._verifier_observer.fire_if_not_fired(verifier)
 
     def got_versions(self, versions):
+        ###XXX _could_ notify _status() via here .. but need "someone"
+        ###to hold "the current status" and that's _RC so far
         self._version_observer.fire_if_not_fired(versions)
 
     def received(self, plaintext):
@@ -263,7 +292,8 @@ def create(
         timing=None,
         stderr=sys.stderr,
         _eventual_queue=None,
-        _enable_dilate=False):
+        _enable_dilate=False,
+        on_status_update=None):
     timing = timing or DebugTiming()
     side = bytes_to_hexstr(os.urandom(5))
     journal = journal or ImmediateJournal()
@@ -272,7 +302,7 @@ def create(
     if delegate:
         w = _DelegatedWormhole(delegate)
     else:
-        w = _DeferredWormhole(reactor, eq, _enable_dilate=_enable_dilate)
+        w = _DeferredWormhole(reactor, eq, _enable_dilate=_enable_dilate, _on_status_update=on_status_update)
     # this indicates Wormhole capabilities
     wormhole_versions = {
         "can-dilate": DILATION_VERSIONS,